+++ tests/split_commit.rs
@@ -17,6 +17,8 @@ mod common;
 
 #[path = "split_commit/dirty_state.rs"]
 mod dirty_state;
+#[path = "split_commit/hunkless.rs"]
+mod hunkless;
 #[path = "split_commit/multi_path.rs"]
 mod multi_path;
 #[path = "split_commit/out_files.rs"]
@@ -33,3 +35,5 @@ mod per_hunk_group;
 mod rename;
 #[path = "split_commit/root_commit.rs"]
 mod root_commit;
+#[path = "split_commit/type_change.rs"]
+mod type_change;